location-data-mcp-server/pyproject.toml
[project]
name = "location-data-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.45.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.0.1",
    "botocore>=1.35.19",
    "boto3>=1.35.19",
    "starlette>=0.38.5",
    "uvicorn>=0.31.0",
    "aws-lambda-powertools==3.12.0",
    "awslabs.mcp_lambda_handler>=0.1.6",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
]
[tool.uv.pip]
target = ".aws-sam/build/LocationDataMCPFunction"